Re-factor futures into a hierarchy; add tests.
