+++ pubsub/google/cloud/pubsub_v1/exceptions.py
@@ -14,15 +14,9 @@
 
 from __future__ import absolute_import
 
-from google.api_core.exceptions import GoogleAPICallError
-from google.cloud.pubsub_v1.exceptions import TimeoutError
-
-
-class PublishError(GoogleAPICallError):
-    pass
+from concurrent.futures import TimeoutError
 
 
 __all__ = (
-    'PublishError',
     'TimeoutError',
 )
